refactor: split app boot modules
